Add new nodes and predicate types
